♻️ refactor: extract raceAbort/resolveWithAbort into shared abort.ts
Both helpers were private to builders.ts but evaluator.ts now needs
raceAbort too. Moved both to src/abort.ts and updated builders and
evaluator to import from there.
